requirements.txt
PySide6
Pillow
rembg
onnxruntime
streamlit
